refactor: Separation of concerns: NodeScene contains/creates/edits nodes. Nodz is the view and only handles view manipulations. API points are re-routed to Nodz.
