+++ nodz_demo.py
@@ -1,6 +1,7 @@
 from typing import Any
 from qtpy import QtCore, QtGui, QtWidgets
-import nodz.core as core
+import nodz.view as view
+import nodz.items as items
 from nodz.utils import nlog
 
 try:
@@ -14,109 +15,106 @@ except BaseException:
 ######################################################################
 
 
-class TestNode(core.NodeItem):
-    def __init__(
-        self,
-        name: str,
-        alternate: bool,
-        preset: str,
-        config: dict,
-        help: str = "",
-    ) -> None:
-        super().__init__(name, alternate, preset, config)
-        self.help = help if help else "This is a subclassed node."
+# class TestNode(items.NodeItem):
+#     def __init__(
+#         self,
+#         name: str,
+#         alternate: bool,
+#         preset: str,
+#         config: dict,
+#         help: str = "",
+#     ) -> None:
+#         super().__init__(name, alternate, preset, config)
+#         self.help = help if help else "This is a subclassed node."
 
-    def to_dict(self) -> dict:
-        d = super().to_dict()
-        d["help"] = self.help
-        return d
+#     def to_dict(self) -> dict:
+#         d = super().to_dict()
+#         d["help"] = self.help
+#         return d
 
-    def configure_from_dict(self, d: dict) -> None:
-        if d and "help" in d:
-            self.help = d["help"]
+#     def configure_from_dict(self, d: dict) -> None:
+#         if d and "help" in d:
+#             self.help = d["help"]
 
-    def paint_attr_label(
-        self,
-        attr: str,
-        painter: QtGui.QPainter,
-        rect: QtCore.QRect,
-        align_flag: QtCore.Qt.AlignmentFlag = QtCore.Qt.AlignmentFlag.AlignVCenter,
-    ):
-        attr_data = self.attrs_data[attr]
+#     def paint_attr_label(
+#         self,
+#         attr: str,
+#         painter: QtGui.QPainter,
+#         rect: QtCore.QRect,
+#         align_flag: QtCore.Qt.AlignmentFlag = QtCore.Qt.AlignmentFlag.AlignVCenter,
+#     ):
+#         attr_data = self.attrs_data[attr]
 
-        align_flag = (
-            QtCore.Qt.AlignmentFlag.AlignVCenter
-            if attr_data["socket"] and not attr_data["plug"]
-            else QtCore.Qt.AlignmentFlag.AlignVCenter
-            | QtCore.Qt.AlignmentFlag.AlignRight
-            if attr_data["plug"] and not attr_data["socket"]
-            else QtCore.Qt.AlignmentFlag.AlignCenter
-        )
-        super().paint_attr_label(
-            attr,
-            painter,
-            rect,
-            align_flag=align_flag,
-        )
-
-
-class TestPlug(core.PlugItem):
-    def __init__(
-        self,
-        parent: QtWidgets.QGraphicsItem,
-        attribute: str,
-        index: int,
-        preset: str,
-        data_type: Any,
-        max_connections: int,
-        help: str = "",
-    ) -> None:
-        super().__init__(
-            parent, attribute, index, preset, data_type, max_connections
-        )
-        self.help = help if help else "This is a subclassed plug."
-
-    def to_dict(self) -> dict:
-        d = super().to_dict()
-        d["help"] = self.help
-        return d
-
-    def configure_from_dict(self, d: dict) -> None:
-        if d and "help" in d:
-            self.help = d["help"]
+#         align_flag = (
+#             QtCore.Qt.AlignmentFlag.AlignVCenter
+#             if attr_data["socket"] and not attr_data["plug"]
+#             else QtCore.Qt.AlignmentFlag.AlignVCenter
+#             | QtCore.Qt.AlignmentFlag.AlignRight
+#             if attr_data["plug"] and not attr_data["socket"]
+#             else QtCore.Qt.AlignmentFlag.AlignCenter
+#         )
+#         super().paint_attr_label(
+#             attr,
+#             painter,
+#             rect,
+#             align_flag=align_flag,
+#         )
 
 
-class TestSocket(core.SocketItem):
-    def __init__(
-        self,
-        parent: QtWidgets.QGraphicsItem,
-        attribute: str,
-        index: int,
-        preset: str,
-        data_type: Any,
-        max_connections: int,
-        help: str = "",
-    ) -> None:
-        super().__init__(
-            parent, attribute, index, preset, data_type, max_connections
-        )
-        self.help = help if help else "This is a subclassed socket."
+# class TestPlug(items.PlugItem):
+#     def __init__(
+#         self,
+#         parent: QtWidgets.QGraphicsItem,
+#         attribute: str,
+#         index: int,
+#         preset: str,
+#         data_type: Any,
+#         max_connections: int,
+#         help: str = "",
+#     ) -> None:
+#         super().__init__(
+#             parent, attribute, index, preset, data_type, max_connections
+#         )
+#         self.help = help if help else "This is a subclassed plug."
 
-    def to_dict(self) -> dict:
-        d = super().to_dict()
-        d["help"] = self.help
-        return d
+#     def to_dict(self) -> dict:
+#         d = super().to_dict()
+#         d["help"] = self.help
+#         return d
 
-    def configure_from_dict(self, d: dict) -> None:
-        if d and "help" in d:
-            self.help = d["help"]
+#     def configure_from_dict(self, d: dict) -> None:
+#         if d and "help" in d:
+#             self.help = d["help"]
 
 
-nodz = core.Nodz(
+# class TestSocket(items.SocketItem):
+#     def __init__(
+#         self,
+#         parent: QtWidgets.QGraphicsItem,
+#         attribute: str,
+#         index: int,
+#         preset: str,
+#         data_type: Any,
+#         max_connections: int,
+#         help: str = "",
+#     ) -> None:
+#         super().__init__(
+#             parent, attribute, index, preset, data_type, max_connections
+#         )
+#         self.help = help if help else "This is a subclassed socket."
+
+#     def to_dict(self) -> dict:
+#         d = super().to_dict()
+#         d["help"] = self.help
+#         return d
+
+#     def configure_from_dict(self, d: dict) -> None:
+#         if d and "help" in d:
+#             self.help = d["help"]
+
+
+nodz = view.Nodz(
     None,
-    nodeitem_cls=TestNode,
-    plugitem_cls=TestPlug,
-    socketitem_cls=TestSocket,
 )
 # nodz.loadConfig(filePath='')
 nodz.initialize()
@@ -129,48 +127,48 @@ nodz.show()
 
 
 # Nodes
-@QtCore.Slot(str)
+@QtCore.Slot(str)  # type: ignore
 def on_nodeCreated(nodeName):
     nlog.info(f"node created : {nodeName}")
 
 
-@QtCore.Slot(str)
+@QtCore.Slot(str)  # type: ignore
 def on_nodeDeleted(nodeName):
     nlog.info(f"node deleted : {nodeName}")
 
 
-@QtCore.Slot(str, str)
+@QtCore.Slot(str, str)  # type: ignore
 def on_nodeEdited(nodeName, newName):
     nlog.info(f"node edited : {nodeName}, new name : {newName}")
 
 
-@QtCore.Slot(str)
+@QtCore.Slot(str)  # type: ignore
 def on_nodeSelected(nodesName):
     nlog.info(f"node selected : {nodesName}")
 
 
-@QtCore.Slot(str, object)
+@QtCore.Slot(str, object)  # type: ignore
 def on_nodeMoved(nodeName, nodePos):
     nlog.info(f"node {nodeName} moved to {nodePos}")
 
 
-@QtCore.Slot(str)
+@QtCore.Slot(str)  # type: ignore
 def on_nodeDoubleClick(nodeName):
     nlog.info(f"double click on node : {nodeName}")
 
 
 # Attrs
-@QtCore.Slot(str, int)
+@QtCore.Slot(str, int)  # type: ignore
 def on_attrCreated(nodeName, attrId):
     nlog.info(f"attr created : {nodeName} at index : {attrId}")
 
 
-@QtCore.Slot(str, int)
+@QtCore.Slot(str, int)  # type: ignore
 def on_attrDeleted(nodeName, attrId):
     nlog.info(f"attr Deleted : {nodeName} at old index : {attrId}")
 
 
-@QtCore.Slot(str, int, int)
+@QtCore.Slot(str, int, int)  # type: ignore
 def on_attrEdited(nodeName, oldId, newId):
     nlog.info(
         f"attr Edited : {nodeName} at old index : {oldId}, new index : {newId}"
@@ -178,7 +176,7 @@ def on_attrEdited(nodeName, oldId, newId):
 
 
 # Connections
-@QtCore.Slot(str, str, str, str)
+@QtCore.Slot(str, str, str, str)  # type: ignore
 def on_connected(srcNodeName, srcPlugName, destNodeName, dstSocketName):
     nlog.info(
         f'connected src: "{srcNodeName}" at "{srcPlugName}" to dst: '
@@ -186,7 +184,7 @@ def on_connected(srcNodeName, srcPlugName, destNodeName, dstSocketName):
     )
 
 
-@QtCore.Slot(str, str, str, str)
+@QtCore.Slot(str, str, str, str)  # type: ignore
 def on_disconnected(srcNodeName, srcPlugName, destNodeName, dstSocketName):
     nlog.info(
         f'disconnected src: "{srcNodeName}" at "{srcPlugName}" from dst: '
@@ -195,52 +193,58 @@ def on_disconnected(srcNodeName, srcPlugName, destNodeName, dstSocketName):
 
 
 # Graph
-@QtCore.Slot()
+@QtCore.Slot()  # type: ignore
 def on_graphSaved():
     nlog.info("graph saved !")
 
 
-@QtCore.Slot()
+@QtCore.Slot()  # type: ignore
 def on_graphLoaded():
     nlog.info("graph loaded !")
 
 
-@QtCore.Slot()
+@QtCore.Slot()  # type: ignore
 def on_graphCleared():
     nlog.info("graph cleared !")
 
 
-@QtCore.Slot()
+@QtCore.Slot()  # type: ignore
 def on_graphEvaluated():
     nlog.info("graph evaluated !")
 
 
 # Other
-@QtCore.Slot(object)
+_last_key = None
+
+
+@QtCore.Slot(object)  # type: ignore
 def on_keyPressed(key):
-    nlog.info(f"key pressed :  {key}")
+    global _last_key
+    if _last_key != key:
+        nlog.info(f"key pressed :  {key}")
+        _last_key = key
 
 
-nodz.signal_NodeCreated.connect(on_nodeCreated)
-nodz.signal_NodeDeleted.connect(on_nodeDeleted)
-nodz.signal_NodeEdited.connect(on_nodeEdited)
-nodz.signal_NodeSelected.connect(on_nodeSelected)
-nodz.signal_NodeMoved.connect(on_nodeMoved)
-nodz.signal_NodeDoubleClicked.connect(on_nodeDoubleClick)
+nodz._scene.signal_NodeCreated.connect(on_nodeCreated)
+nodz._scene.signal_NodeDeleted.connect(on_nodeDeleted)
+nodz._scene.signal_NodeEdited.connect(on_nodeEdited)
+nodz._scene.signal_NodeSelected.connect(on_nodeSelected)
+nodz._scene.signal_NodeMoved.connect(on_nodeMoved)
+nodz._scene.signal_NodeDoubleClicked.connect(on_nodeDoubleClick)
 
-nodz.signal_AttrCreated.connect(on_attrCreated)
-nodz.signal_AttrDeleted.connect(on_attrDeleted)
-nodz.signal_AttrEdited.connect(on_attrEdited)
+nodz._scene.signal_AttrCreated.connect(on_attrCreated)
+nodz._scene.signal_AttrDeleted.connect(on_attrDeleted)
+nodz._scene.signal_AttrEdited.connect(on_attrEdited)
 
-nodz.signal_PlugConnected.connect(on_connected)
-nodz.signal_SocketConnected.connect(on_connected)
-nodz.signal_PlugDisconnected.connect(on_disconnected)
-nodz.signal_SocketDisconnected.connect(on_disconnected)
+nodz._scene.signal_PlugConnected.connect(on_connected)
+nodz._scene.signal_SocketConnected.connect(on_connected)
+nodz._scene.signal_PlugDisconnected.connect(on_disconnected)
+nodz._scene.signal_SocketDisconnected.connect(on_disconnected)
 
-nodz.signal_GraphSaved.connect(on_graphSaved)
-nodz.signal_GraphLoaded.connect(on_graphLoaded)
-nodz.signal_GraphCleared.connect(on_graphCleared)
-nodz.signal_GraphEvaluated.connect(on_graphEvaluated)
+nodz._scene.signal_GraphSaved.connect(on_graphSaved)
+nodz._scene.signal_GraphLoaded.connect(on_graphLoaded)
+nodz._scene.signal_GraphCleared.connect(on_graphCleared)
+nodz._scene.signal_GraphEvaluated.connect(on_graphEvaluated)
 
 nodz.signal_KeyPressed.connect(on_keyPressed)
 
@@ -253,7 +257,7 @@ nodz.signal_KeyPressed.connect(on_keyPressed)
 nodeA = nodz.create_node(name="nodeA", preset="node_preset_1", position=None)
 
 nodz.create_attribute(
-    node=nodeA,
+    nodeA,
     name="Aattr1",
     index=-1,
     preset="attr_preset_1",
@@ -263,7 +267,7 @@ nodz.create_attribute(
 )
 
 nodz.create_attribute(
-    node=nodeA,
+    nodeA,
     name="Aattr2",
     index=-1,
     preset="attr_preset_1",
@@ -273,7 +277,7 @@ nodz.create_attribute(
 )
 
 nodz.create_attribute(
-    node=nodeA,
+    nodeA,
     name="Aattr3",
     index=-1,
     preset="attr_preset_2",
@@ -283,7 +287,7 @@ nodz.create_attribute(
 )
 
 nodz.create_attribute(
-    node=nodeA,
+    nodeA,
     name="Aattr4",
     index=-1,
     preset="attr_preset_2",
@@ -293,7 +297,7 @@ nodz.create_attribute(
 )
 
 nodz.create_attribute(
-    node=nodeA,
+    nodeA,
     name="Aattr5",
     index=-1,
     preset="attr_preset_3",
@@ -305,7 +309,7 @@ nodz.create_attribute(
 )
 
 nodz.create_attribute(
-    node=nodeA,
+    nodeA,
     name="Aattr6",
     index=-1,
     preset="attr_preset_3",
@@ -321,7 +325,7 @@ nodz.create_attribute(
 nodeB = nodz.create_node(name="nodeB", preset="node_preset_1")
 
 nodz.create_attribute(
-    node=nodeB,
+    nodeB,
     name="Battr1",
     index=-1,
     preset="attr_preset_1",
@@ -331,7 +335,7 @@ nodz.create_attribute(
 )
 
 nodz.create_attribute(
-    node=nodeB,
+    nodeB,
     name="Battr2",
     index=-1,
     preset="attr_preset_1",
@@ -341,7 +345,7 @@ nodz.create_attribute(
 )
 
 nodz.create_attribute(
-    node=nodeB,
+    nodeB,
     name="Battr3",
     index=-1,
     preset="attr_preset_2",
@@ -351,7 +355,7 @@ nodz.create_attribute(
 )
 
 nodz.create_attribute(
-    node=nodeB,
+    nodeB,
     name="Battr4",
     index=-1,
     preset="attr_preset_3",
@@ -367,7 +371,7 @@ nodz.create_attribute(
 nodeC = nodz.create_node(name="nodeC", preset="node_preset_1")
 
 nodz.create_attribute(
-    node=nodeC,
+    nodeC,
     name="Cattr1",
     index=-1,
     preset="attr_preset_1",
@@ -377,7 +381,7 @@ nodz.create_attribute(
 )
 
 nodz.create_attribute(
-    node=nodeC,
+    nodeC,
     name="Cattr2",
     index=-1,
     preset="attr_preset_1",
@@ -387,7 +391,7 @@ nodz.create_attribute(
 )
 
 nodz.create_attribute(
-    node=nodeC,
+    nodeC,
     name="Cattr3",
     index=-1,
     preset="attr_preset_1",
@@ -397,7 +401,7 @@ nodz.create_attribute(
 )
 
 nodz.create_attribute(
-    node=nodeC,
+    nodeC,
     name="Cattr4",
     index=-1,
     preset="attr_preset_2",
@@ -407,7 +411,7 @@ nodz.create_attribute(
 )
 
 nodz.create_attribute(
-    node=nodeC,
+    nodeC,
     name="Cattr5",
     index=-1,
     preset="attr_preset_2",
@@ -417,7 +421,7 @@ nodz.create_attribute(
 )
 
 nodz.create_attribute(
-    node=nodeC,
+    nodeC,
     name="Cattr6",
     index=-1,
     preset="attr_preset_3",
@@ -427,7 +431,7 @@ nodz.create_attribute(
 )
 
 nodz.create_attribute(
-    node=nodeC,
+    nodeC,
     name="Cattr7",
     index=-1,
     preset="attr_preset_3",
@@ -437,7 +441,7 @@ nodz.create_attribute(
 )
 
 nodz.create_attribute(
-    node=nodeC,
+    nodeC,
     name="Cattr8",
     index=-1,
     preset="attr_preset_3",
@@ -450,7 +454,7 @@ nodz.create_attribute(
 nodeD = nodz.create_node(name="nodeD", preset="node_preset_1")
 
 nodz.create_attribute(
-    node=nodeD,
+    nodeD,
     name="Dattr1",
     index=-1,
     preset="attr_preset_3",
@@ -460,7 +464,7 @@ nodz.create_attribute(
 )
 
 nodz.create_attribute(
-    node=nodeD,
+    nodeD,
     name="Dattr2",
     index=-1,
     preset="attr_preset_3",
@@ -473,7 +477,7 @@ nodz.create_attribute(
 nodeE = nodz.create_node(name="nodeE", preset="node_preset_1")
 
 nodz.create_attribute(
-    node=nodeE,
+    nodeE,
     name="Eattr1",
     index=-1,
     preset="attr_preset_1",
@@ -483,7 +487,7 @@ nodz.create_attribute(
 )
 
 nodz.create_attribute(
-    node=nodeE,
+    nodeE,
     name="Eattr2",
     index=-1,
     preset="attr_preset_2",
@@ -493,7 +497,7 @@ nodz.create_attribute(
 )
 
 nodz.create_attribute(
-    node=nodeE,
+    nodeE,
     name="Eattr3",
     index=-1,
     preset="attr_preset_2",
@@ -506,7 +510,7 @@ nodz.create_attribute(
 nodeF = nodz.create_node(name="nodeF", preset="node_preset_1")
 
 nodz.create_attribute(
-    node=nodeF,
+    nodeF,
     name="Fattr1",
     index=-1,
     preset="attr_preset_1",
@@ -516,7 +520,7 @@ nodz.create_attribute(
 )
 
 nodz.create_attribute(
-    node=nodeF,
+    nodeF,
     name="Fattr2",
     index=-1,
     preset="attr_preset_2",
@@ -526,7 +530,7 @@ nodz.create_attribute(
 )
 
 nodz.create_attribute(
-    node=nodeF,
+    nodeF,
     name="Fattr3",
     index=-1,
     preset="attr_preset_2",
@@ -548,20 +552,18 @@ nodz.create_connection("nodeD", "Dattr2", "nodeA", "Aattr6")
 nodz.create_connection("nodeE", "Eattr1", "nodeF", "Fattr2")
 
 # Attributes Edition
-nodz.edit_attribute(node=nodeC, index=0, new_name=None, new_index=-1)
-nodz.edit_attribute(
-    node=nodeC, index=-1, new_name="NewAttrName", new_index=None
-)
+nodz.edit_attribute(nodeC, index=0, new_name=None, new_index=-1)
+nodz.edit_attribute(nodeC, index=-1, new_name="NewAttrName", new_index=None)
 
 # Attributes Deletion
-nodz.delete_attribute(node=nodeC, index=-1)
+nodz.delete_attribute(nodeC, index=-1)
 
 
 # Nodes Edition
-nodz.edit_node(node=nodeC, new_name="newNodeName")
+nodeC = nodz.edit_node(nodeC, new_name="newNodeName")
 
 # Nodes Deletion
-nodz.delete_node(node=nodeC)
+nodz.delete_node(nodeC)
 
 
 # Graph
